+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+* **New Resource:** `gigahost_dns_record`
 * **New Resource:** `gigahost_dns_zone`
 * **New Data Source:** `gigahost_account`
 * **New Data Source:** `gigahost_dns_zones`